# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaClone API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor de API Principal"
        }
    ],
    "paths": {
...
        "/posts/{postId}/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Listar comentários de um post",
                "description": "Retorna todos os comentários associados a um post específico",
                "operationId": "48be78ad2417cc7a2b2c87765a6cb6d0",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de comentários retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "post_id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "Ótima foto!"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-05-22T10:30:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-05-22T10:30:00Z"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Maria Silva"
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "maria_silva"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "/storage/avatars/maria.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Criar comentário em um post",
                "description": "Cria um novo comentário associado a um post (requer autenticação)",
                "operationId": "ce8360d198d910f414bf4e4af039712a",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do comentário a ser criado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Adorei essa foto!",
                                        "maxLength": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comentário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "post_id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Adorei essa foto!"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:35:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:35:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
        "/comments/{commentId}": {
            "put": {
                "tags": [
                    "Comment"
                ],
                "summary": "Atualizar comentário",
                "description": "Atualiza o conteúdo de um comentário existente (apenas o autor pode atualizar)",
                "operationId": "9ab14b62a644898bdd7a1d441c9f9b6c",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID do comentário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Novo conteúdo do comentário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Adorei essa foto! Muito legal.",
                                        "maxLength": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comentário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "post_id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Adorei essa foto! Muito legal."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:35:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:40:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso negado (apenas o autor pode atualizar)"
                    },
                    "404": {
                        "description": "Comentário não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Deletar comentário",
                "description": "Remove um comentário existente (apenas o autor ou admin podem deletar)",
                "operationId": "20c6cd5be2b4338d3ae580f108fab7a4",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID do comentário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comentário removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Removido com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso negado (apenas o autor pode deletar)"
                    },
                    "404": {
                        "description": "Comentário não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}